Use 16 bit int for ShortType

+++ Rich/RichUtils/RichUtils/RichDAQDefinitions.h
@@ -35,7 +35,7 @@ namespace Rich
     using LongType  = std::uint32_t;
 
     /// DAQ short type definition
-    using ShortType = std::uint32_t;
+    using ShortType = std::uint16_t;
 
     //---------------------------------------------------------------------------------
